added guide code for basic usage, Usage with Crawler and Explicit usage with Crawler
