+++ docs/guides/code/result_storage_dataset_basic.py
@@ -1,29 +1,21 @@
 import asyncio
 
-from crawlee.beautifulsoup_crawler import BeautifulSoupCrawler, BeautifulSoupCrawlingContext
+# Import the Dataset class to manage data storage
 from crawlee.storages import Dataset
 
 
 async def main() -> None:
-    crawler = BeautifulSoupCrawler()
-    # Open dataset manually using asynchronous constructor open().
+    # Open a default dataset using the asynchronous constructor.
+    # This dataset is used to store data in the Crawlee storage.
     dataset = await Dataset.open()
 
-    # A decorator that registers a default handler, invoked for requests without a label or with an unrecognized label.
-    @crawler.router.default_handler
-    async def request_handler(context: BeautifulSoupCrawlingContext) -> None:
-        context.log.info(f'Processing {context.request.url} ...')
+    # Push a single row of data (a dictionary) to the named dataset.
+    await dataset.push_data({'foo': 'bar'})
 
-        # Extract data from the page.
-        data = {
-            'url': context.request.url,
-            'title': context.soup.title.string if context.soup.title else None,
-            'html': str(context.soup)[:1000],
-        }
-
-        # Push the extracted data to the dataset.
-        await dataset.push_data(data)
+    # Push multiple rows of data (a list of dictionaries) to the named dataset.
+    await dataset.push_data([{'foo': 'bar2', 'col2': 'val2'}, {'col3': 123}])
 
 
+# Run the main function in an asynchronous event loop
 if __name__ == '__main__':
     asyncio.run(main())
